Support message option in all validators
