+++ lib/matchers/validations/confirmation_of.rb
@@ -2,15 +2,8 @@ module Mongoid
   module Matchers
     module Validations
       class ValidateConfirmationOfMatcher < HaveValidationMatcher
-        include WithMessage
-
         def initialize(name)
           super(name, :confirmation)
-        end
-
-        def with_message(message)
-          @expected_message = message
-          self
         end
       end
