+++ lib/matchers/validations/custom_validation_of.rb
@@ -2,7 +2,6 @@ module Mongoid
   module Matchers
     module Validations
       class ValidateWithCustomValidatorMatcher < HaveValidationMatcher
-        include WithMessage
         def initialize(field)
           super(field, :custom)
         end
@@ -15,7 +14,6 @@ module Mongoid
         def matches?(actual)
           return false unless (@result = super(actual))
           check_custom_validator if @custom_validator
-          check_expected_message if @expected_message
 
           @result
         end
@@ -23,7 +21,6 @@ module Mongoid
         def description
           options_desc = []
           options_desc << " with custom validator #{@custom_validator.name}" if @validator
-          options_desc << " with message '#{@expected_message}'" if @expected_message
           "validate field #{@field.inspect}" << options_desc.to_sentence
         end
