+++ lib/mongoid/rspec.rb
@@ -13,7 +13,6 @@ require 'matchers/associations'
 require 'matchers/allow_mass_assignment'
 require 'matchers/accept_nested_attributes'
 require 'matchers/validations'
-require 'matchers/validations/with_message'
 require 'matchers/validations/associated'
 require 'matchers/validations/confirmation_of'
 require 'matchers/validations/exclusion_of'
