+++ spec/models/message.rb
@@ -1,0 +1,11 @@
+class Message
+  include Mongoid::Document
+
+  field :identifier
+  field :from
+  field :to
+
+  validates :identifier, uniqueness: { message: 'uniqueness' }
+  validates :from, presence: { message: 'required' }
+  validates :to, format: { with: /[a-z]+/, message: 'format' }
+end
